refactor(movement): move light client types
refactor(movement): move types to crate

fix(movement): light-client types

refactor(movement): fix type errors

fix(movement): proto impl gen
