+++ lib/unionlabs/src/ibc/lightclients.rs
@@ -1,4 +1,3 @@
 pub mod ethereum;
 pub mod evm_in_cosmos;
-pub mod movement;
 pub mod wasm;
