+++ light-clients/movement/ics08-movement/src/contract.rs
@@ -3,12 +3,13 @@ use ics008_wasm_client::{
     storage_utils::{save_proto_client_state, save_proto_consensus_state},
     CustomQueryOf, InstantiateMsg,
 };
+use movement_light_client_types::ClientState;
 use protos::ibc::lightclients::wasm::v1::{
     ClientState as ProtoClientState, ConsensusState as ProtoConsensusState,
 };
 use unionlabs::{
     encoding::{DecodeAs, Proto},
-    ibc::{core::client::height::Height, lightclients::movement::client_state::ClientState},
+    ibc::core::client::height::Height,
 };
 
 use crate::{client::MovementLightClient, errors::Error};
