+++ light-clients/movement/ics08-movement/types/src/lib.rs
@@ -1,0 +1,5 @@
+pub mod client_state;
+pub mod consensus_state;
+pub mod header;
+
+pub use crate::{client_state::ClientState, consensus_state::ConsensusState, header::Header};
