+++ voyager/modules/client/movement/Cargo.toml
@@ -4,20 +4,21 @@ name    = "voyager-client-module-movement"
 version = "0.1.0"
 
 [dependencies]
-chain-utils        = { workspace = true }
-enumorph           = { workspace = true }
-futures            = { workspace = true }
-jsonrpsee          = { workspace = true, features = ["macros", "server", "tracing"] }
-macros             = { workspace = true }
-prost              = { workspace = true }
-protos             = { workspace = true }
-serde              = { workspace = true, features = ["derive"] }
-serde-utils        = { workspace = true }
-serde_json         = { workspace = true }
-thiserror          = { workspace = true }
-tokio              = { workspace = true }
-tracing            = { workspace = true }
-tracing-subscriber = { workspace = true }
-unionlabs          = { workspace = true }
-voyager-message    = { workspace = true }
-voyager-vm         = { workspace = true }
+chain-utils                 = { workspace = true }
+enumorph                    = { workspace = true }
+futures                     = { workspace = true }
+jsonrpsee                   = { workspace = true, features = ["macros", "server", "tracing"] }
+macros                      = { workspace = true }
+movement-light-client-types = { workspace = true, features = ["serde", "proto"] }
+prost                       = { workspace = true }
+protos                      = { workspace = true }
+serde                       = { workspace = true, features = ["derive"] }
+serde-utils                 = { workspace = true }
+serde_json                  = { workspace = true }
+thiserror                   = { workspace = true }
+tokio                       = { workspace = true }
+tracing                     = { workspace = true }
+tracing-subscriber          = { workspace = true }
+unionlabs                   = { workspace = true }
+voyager-message             = { workspace = true }
+voyager-vm                  = { workspace = true }
