+++ voyager/plugins/client-update/movement/Cargo.toml
@@ -4,28 +4,29 @@ name    = "voyager-client-update-plugin-movement"
 version = "0.1.0"
 
 [dependencies]
-aptos-move-ibc     = { workspace = true }
-aptos-rest-client  = { workspace = true }
-aptos-types        = { workspace = true }
-bitvec             = { workspace = true }
-chain-utils        = { workspace = true }
-enumorph           = { workspace = true }
-ethers             = { workspace = true, features = ["rustls", "ws"] }
-futures            = { workspace = true }
-jsonrpsee          = { workspace = true, features = ["macros", "server", "tracing"] }
-macros             = { workspace = true }
-move-core-types    = { workspace = true }
-num-bigint         = { workspace = true }
-prost              = { workspace = true }
-protos             = { workspace = true }
-reqwest            = { workspace = true, features = ["json"] }
-serde              = { workspace = true, features = ["derive"] }
-serde_json         = { workspace = true }
-subset-of          = { workspace = true }
-thiserror          = { workspace = true }
-tokio              = { workspace = true }
-tracing            = { workspace = true }
-tracing-subscriber = { workspace = true }
-unionlabs          = { workspace = true }
-voyager-message    = { workspace = true }
-voyager-vm         = { workspace = true }
+aptos-move-ibc              = { workspace = true }
+aptos-rest-client           = { workspace = true }
+aptos-types                 = { workspace = true }
+bitvec                      = { workspace = true }
+chain-utils                 = { workspace = true }
+enumorph                    = { workspace = true }
+ethers                      = { workspace = true, features = ["rustls", "ws"] }
+futures                     = { workspace = true }
+jsonrpsee                   = { workspace = true, features = ["macros", "server", "tracing"] }
+macros                      = { workspace = true }
+move-core-types             = { workspace = true }
+movement-light-client-types = { workspace = true, features = ["serde"] }
+num-bigint                  = { workspace = true }
+prost                       = { workspace = true }
+protos                      = { workspace = true }
+reqwest                     = { workspace = true, features = ["json"] }
+serde                       = { workspace = true, features = ["derive"] }
+serde_json                  = { workspace = true }
+subset-of                   = { workspace = true }
+thiserror                   = { workspace = true }
+tokio                       = { workspace = true }
+tracing                     = { workspace = true }
+tracing-subscriber          = { workspace = true }
+unionlabs                   = { workspace = true }
+voyager-message             = { workspace = true }
+voyager-vm                  = { workspace = true }
